requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary
boto3==1.34.156
django-cors-headers==4.4.0
django-cron==0.6.0
python-dotenv==1.0.1
PyJWT==2.9.0
